Docs: fix `UFix64`, `Fix64` and `Address` in types tables (#1073)
It seems that the `UFix64` and `Fix64` type arguments are not correctly
represented if you pass them as arguments with the `t.UFix256` or
`t.Fix256` type respectively.

After a quick look in `types.js` it seems that the library doesn't even
know about those types. Passing in `t.UFix64` and `t.Fix64` as a type
seems to work just fine.

https://github.com/onflow/fcl-js/blob/master/packages/types/src/types.js

It also seems that an `Address` typed argument needs to be passed in
with the `t.Address` type instead of the `t.String` type as suggested
in the docs.

This PR additionally fixes the formatting/styling for the `path` type
at the very end of the table in the docs.

Co-authored-by: Chase Fleming <chasefleming@users.noreply.github.com>