Docs: fix reference to undefined `args` types (#1072)
The two examples show code that rely on the `t` object to be available
in order to pass in a type reference to the `args` builder.

But in the shown examples the `args` are constructed in a different way
where the `t` object is not passed in as a function argument and
therefore results in a `ReferenceError`.

Co-authored-by: Chase Fleming <chasefleming@users.noreply.github.com>